fix: modified load of passport settings which tracks if player connected via pkce
chore: moved data types and requests data definitions from headers
